+++ src/components/my_page/mini-profile.tsx
