+++ src/components/my_page/pghd-record.tsx
