+++ src/components/sign_up/nick-name.tsx
