+++ src/components/sign_up/password.tsx
