+++ src/components/sign_up/terms.tsx
